버스 정류장 명칭뿐만 아니라 번호로 검색이 되도록 동작 수정

+++ app/src/main/java/com/custom/transportation/repository/local/Converts.kt
@@ -12,7 +12,7 @@ class Converts {
         data ?: return null
         val splits = data.split('/')
         return if(splits.size != 5)  null
-        else BusStopData(splits[0].toInt(), splits[1].toInt(), splits[2], splits[3].toFloat(), splits[4].toFloat())
+        else BusStopData(splits[0], splits[1], splits[2], splits[3].toFloat(), splits[4].toFloat())
     }
 
     @TypeConverter
